Refactor: Reorganize automate-related files and update imports
- Move and rename automation-related files to improve project structure
- Update import paths for automate actions, slices, and components
- Remove deprecated and commented-out code
- Simplify component imports and remove unused dependencies
