+++ app/controllers/automate.js
