--- app/utils/addon.json
@@ -1,1 +1,0 @@
-{}
