+++ src/actions/automate.js
@@ -1,8 +1,49 @@
 import apiBackend from "@src/apiBackend.js";
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
+
+// -------------------------------
+export const goAutomate = createAsyncThunk(
+  "automate/goAutomate",
+  async ({ userId, email, password, query, tokenGpt, /*logs,*/ ftpData, file }) => {
+    try {
+      console.log("EMAIL FETCH REQUEST:", { userId, email, query, ftpData });
+      const token = localStorage.getItem("token");
+      const formData = new FormData();
+      
+      formData.append('userId', userId);
+      formData.append('email', email);
+      formData.append('password', password);
+      formData.append('query', query);
+      formData.append('tokenGpt', tokenGpt);
+      formData.append('ftpData', JSON.stringify(ftpData));
+      if (file) {
+        formData.append('file', file);
+      }
+
+      const res = await apiBackend.post(
+        "/automate/go",
+        formData,
+        {
+          headers: {
+            "Content-Type": "multipart/form-data",
+            Authorization: `Bearer ${token}`,
+          },
+        }
+      );
+
+      console.log("EMAIL FETCH RESPONSE:", res.data);
+      return res.data; // Return the email data
+    } catch (error) {
+      console.error("Error in getEmailsByQuery action:", error);
+      return rejectWithValue(error.response?.data || "Failed to fetch emails");
+    }
+  }
+);
+
+
 export const createAutomation = createAsyncThunk(
-  "automations/createAutomation",
+  "automate/createAutomation",
   async ({ userId, email, automationData }) => {
     console.log("data from createAutomation", {
       userId,
@@ -12,7 +53,7 @@ export const createAutomation = createAsyncThunk(
     try {
       const token = localStorage.getItem("token");
       const res = await apiBackend.post(
-        `/automations/createAutomation`,
+        `/automate/createAutomation`,
         { userId, email, automationData },
         {
           headers: {
@@ -31,12 +72,12 @@ export const createAutomation = createAsyncThunk(
 );
 
 export const getAllUserAutomations = createAsyncThunk(
-  "automations/getAllUserAutomations",
+  "automate/getAllUserAutomations",
   async ({ userId }) => {
     try {
       const token = localStorage.getItem("token");
       const res = await apiBackend.get(
-        `/automations/getAllUserAutomations/${userId}`,
+        `/automate/getAllUserAutomations/${userId}`,
         {
           headers: { Authorization: `Bearer ${token}` },
         }
@@ -54,7 +95,7 @@ export const getAllUserAutomations = createAsyncThunk(
 );
 
 export const updateAutomation = createAsyncThunk(
-  "automations/updateAutomation",
+  "automate/updateAutomation",
   async ({ automationId, toUpdate, userId }) => {
     console.log(
       "Updating automation with ID:",
@@ -65,7 +106,7 @@ export const updateAutomation = createAsyncThunk(
     try {
       const token = localStorage.getItem("token");
       const res = await apiBackend.put(
-        `/automations/updateAutomation/${automationId}`,
+        `/automate/updateAutomation/${automationId}`,
         { ...toUpdate, userId },
         {
           headers: {
@@ -84,13 +125,13 @@ export const updateAutomation = createAsyncThunk(
 );
 
 export const deleteAutomation = createAsyncThunk(
-  "automations/deleteAutomation",
+  "automate/deleteAutomation",
   async ({ automationId, userId }) => {
     console.log("Deleting automation with ID:", automationId);
     try {
       const token = localStorage.getItem("token");
       const res = await apiBackend.delete(
-        `/automations/deleteAutomation/${automationId}`,
+        `/automate/deleteAutomation/${automationId}`,
         {
           headers: {
             Authorization: `Bearer ${token}`,
