+++ src/utils/store.js
@@ -1,7 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
 import emailManagerReducer from "../slices/emailManagerSlices";
 import menuAutomateSlices from "../slices/menuAutomateSlices";
-import automationsSlices from "../slices/automationsSlices";
+import automateSlices from "../slices/automateSlices";
 import scalewaySlices from "../slices/scalewaySlices";
 import clientsSlices from "../slices/clientsSlices";
 import userSlices from "../slices/userSlices";
@@ -12,7 +12,7 @@ const store = configureStore({
   reducer: {
     // emailManager: emailManagerReducer,
     menuAutomate: menuAutomateSlices,
-    automations: automationsSlices,
+    automate: automateSlices,
     scaleway: scalewaySlices,
     clients: clientsSlices,
     user: userSlices,
