+++ src/views/Dashboard/components/Automate/Components/GoogleDriveFormCreateAutomate/GoogleDriveFormCreateAutomate.jsx
@@ -2,12 +2,9 @@ import React, { useState } from "react";
 import TitleFormsComponent from "../../shared/TitleFormsComponent";
 import styles from "../GmailFormCreateAutomate/gmail.module.css";
 import { ReactComponent as DriveIcon } from "../../../../assets/drive.svg";
-import SelectComponent from "../../../SelectComponent/SelectComponent";
 import InputComponent from "../../../InputComponent/InputComponent";
 import SearchSVG from "../../svgs/SearchSVG";
-import CheckboxComponent from "../../shared/CheckboxComponent";
 import OptionsSwitchComponent from "../../../OptionsSwichComponent/OptionsSwitchComponent";
-import LabelSVG from "../../svgs/LabelSVG";
 import TextSVG from "../../svgs/TextSVG";
 import { ReactComponent as GmailIcon } from "../../../../assets/gmail.svg";
 import { ReactComponent as WhatsAppIcon } from "../../../../assets/whatsappIcon.svg";
